refactor(brutus): split test files to match source file organization
