+++ pkg/brutus/brutus_test.go
@@ -15,84 +15,11 @@
 package brutus
 
 import (
-	"context"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 )
-
-func TestIsStandardBanner(t *testing.T) {
-	tests := []struct {
-		name     string
-		protocol string
-		banner   string
-		want     bool
-	}{
-		{
-			name:     "empty banner returns true",
-			protocol: "ssh",
-			banner:   "",
-			want:     true,
-		},
-		{
-			name:     "ssh standard banner",
-			protocol: "ssh",
-			banner:   "SSH-2.0-OpenSSH_8.2",
-			want:     true,
-		},
-		{
-			name:     "ssh custom banner",
-			protocol: "ssh",
-			banner:   "SSH-2.0-MyCustomSSH",
-			want:     false,
-		},
-		{
-			name:     "http protocol with grafana banner should return false",
-			protocol: "http",
-			banner:   "Grafana v8.0.0",
-			want:     false,
-		},
-		{
-			name:     "https protocol with jenkins banner should return false",
-			protocol: "https",
-			banner:   "Jenkins/2.303",
-			want:     false,
-		},
-		{
-			name:     "couchdb protocol with custom banner should return false",
-			protocol: "couchdb",
-			banner:   "CouchDB/3.1.1",
-			want:     false,
-		},
-		{
-			name:     "elasticsearch protocol should return false",
-			protocol: "elasticsearch",
-			banner:   "Elasticsearch 7.10.0",
-			want:     false,
-		},
-		{
-			name:     "influxdb protocol should return false",
-			protocol: "influxdb",
-			banner:   "InfluxDB v2.0",
-			want:     false,
-		},
-		{
-			name:     "unknown non-http protocol returns true",
-			protocol: "unknown",
-			banner:   "some banner",
-			want:     true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := IsStandardBanner(tt.protocol, tt.banner)
-			assert.Equal(t, tt.want, got, "IsStandardBanner(%q, %q) = %v, want %v",
-				tt.protocol, tt.banner, got, tt.want)
-		})
-	}
-}
 
 func TestConfigValidate(t *testing.T) {
 	tests := []struct {
@@ -206,49 +133,5 @@ func TestConfigValidate(t *testing.T) {
 				}
 			}
 		})
-	}
-}
-
-func TestCaptureBanner_EmptyUsernames(t *testing.T) {
-	// Setup: Config with only Credentials (no Usernames), HTTP protocol, LLM enabled
-	cfg := &Config{
-		Target:   "example.com:80",
-		Protocol: "http",
-		Credentials: []Credential{
-			{Username: "admin", Password: "password123"},
-		},
-		Usernames: []string{}, // EMPTY - triggers the bug
-		Timeout:   10 * time.Second,
-		LLMConfig: &LLMConfig{
-			Enabled:  true,
-			Provider: "claude",
-		},
-	}
-
-	// Create mock plugin
-	mockPlugin := &mockHTTPPlugin{}
-
-	// This should NOT panic
-	ctx := context.Background()
-	banner := captureBanner(ctx, cfg, mockPlugin)
-
-	// Should return a valid BannerInfo with empty Banner (not crash)
-	assert.Equal(t, "http", banner.Protocol)
-	assert.Equal(t, "example.com:80", banner.Target)
-	// Banner may be empty, which is fine
-}
-
-type mockHTTPPlugin struct{}
-
-func (m *mockHTTPPlugin) Name() string { return "http" }
-
-func (m *mockHTTPPlugin) Test(ctx context.Context, target, username, password string, timeout time.Duration) *Result {
-	return &Result{
-		Protocol: "http",
-		Target:   target,
-		Username: username,
-		Password: password,
-		Success:  false,
-		Banner:   "HTTP/1.1 401 Unauthorized",
 	}
 }
